Update str_replace.js
The optional 'count' parameter does not return the correct value.
The correct value can be determined by examining the number of elements in the 'split' array.
The count will always be 1 less than this value. 
Note: If the search element is not found split still returns a value of 1 (the whole string) and since 1 - 1 = 0 the current count value remains the same.